Animations
In this commit, we've introduced captivating animations to elevate the user experience when interacting with obstacles in the simulation. Two key animations have been implemented to add a touch of realism and engagement:

Obstacle Scaling on Drag:
Now, when obstacles are dragged, they dynamically scale, providing users with visual feedback as they manipulate the obstacles. Dragging an obstacle upwards increases its size, while dragging downwards reduces it to the original size. This responsive scaling not only adds a playful aspect to obstacle manipulation but also offers a more intuitive feel to the entire interaction process.

Impact Shake Animation:
To intensify the impact of the game, we've incorporated a shake animation for obstacles upon collision with the ball. When the ball collides with an obstacle, the obstacle reacts with a subtle yet noticeable shake effect, creating a sense of impact and enhancing the overall realism of the simulation.